backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
# SaaS Dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.5.0
sendgrid==6.10.0
python-decouple==3.8
supabase==2.0.0
pydantic-settings==2.0.3
alembic==1.12.1
